refactor: extract Sync service into Action classes
Decompose SyncService into SyncGitLab and SyncBitrix24 actions under
Domain/Sync/Actions. Update RunSyncJob and SyncCommand to use actions
directly. Delete SyncService, its interface, and SyncServiceProvider.
