+++ backend/app/Jobs/RunSyncJob.php
@@ -4,11 +4,12 @@ declare(strict_types=1);
 
 namespace App\Jobs;
 
+use App\Domain\Matching\Actions\MatchAllUnmatched;
+use App\Domain\Sync\Actions\SyncBitrix24;
+use App\Domain\Sync\Actions\SyncGitLab;
 use App\Domain\Sync\Enums\SyncStatus;
 use App\Domain\Sync\Enums\SyncStep;
 use App\Domain\Sync\Models\SyncJob;
-use App\Domain\Matching\Actions\MatchAllUnmatched;
-use App\Services\Sync\SyncServiceInterface;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
@@ -34,16 +35,25 @@ class RunSyncJob implements ShouldQueue
     ) {
     }
 
-    public function handle(SyncServiceInterface $syncService, MatchAllUnmatched $matchAllUnmatched): void
+    public function handle(SyncGitLab $syncGitLab, SyncBitrix24 $syncBitrix24, MatchAllUnmatched $matchAllUnmatched): void
     {
         $syncJob = SyncJob::findOrFail($this->syncJobId);
 
         try {
-            if ($this->dateFrom !== null && $this->dateTo !== null) {
-                $this->runResync($syncJob, $syncService, $matchAllUnmatched);
-            } else {
-                $this->runFullSync($syncJob, $syncService, $matchAllUnmatched);
-            }
+            $syncJob->markStep(SyncStep::GitLab);
+            $this->publishProgress(['status' => 'in_progress', 'current_step' => 'gitlab']);
+            $syncGitLab($this->dateFrom, $this->dateTo);
+
+            $syncJob->markStep(SyncStep::Bitrix24);
+            $this->publishProgress(['status' => 'in_progress', 'current_step' => 'bitrix24']);
+            $syncBitrix24();
+
+            $syncJob->markStep(SyncStep::Matching);
+            $this->publishProgress(['status' => 'in_progress', 'current_step' => 'matching']);
+            $matchAllUnmatched();
+
+            $syncJob->markCompleted();
+            $this->publishProgress(['status' => 'success']);
         } catch (\Throwable $e) {
             $syncJob->markFailed($e->getMessage());
             $this->publishProgress(['status' => 'failed', 'error_message' => $e->getMessage()]);
@@ -60,53 +70,6 @@ class RunSyncJob implements ShouldQueue
             $syncJob->markFailed($exception->getMessage());
             $this->publishProgress(['status' => 'failed', 'error_message' => $exception->getMessage()]);
         }
-    }
-
-    private function runFullSync(
-        SyncJob $syncJob,
-        SyncServiceInterface $syncService,
-        MatchAllUnmatched $matchAllUnmatched,
-    ): void {
-        $syncJob->markStep(SyncStep::GitLab);
-        $this->publishProgress(['status' => 'in_progress', 'current_step' => 'gitlab']);
-        $syncService->syncGitLab();
-
-        $syncJob->markStep(SyncStep::Bitrix24);
-        $this->publishProgress(['status' => 'in_progress', 'current_step' => 'bitrix24']);
-        $syncService->syncBitrix24();
-
-        $syncJob->markStep(SyncStep::Matching);
-        $this->publishProgress(['status' => 'in_progress', 'current_step' => 'matching']);
-        $matchAllUnmatched();
-
-        $syncJob->markCompleted();
-        $this->publishProgress(['status' => 'success']);
-    }
-
-    private function runResync(
-        SyncJob $syncJob,
-        SyncServiceInterface $syncService,
-        MatchAllUnmatched $matchAllUnmatched,
-    ): void {
-        /** @var string $dateFrom */
-        $dateFrom = $this->dateFrom;
-        /** @var string $dateTo */
-        $dateTo = $this->dateTo;
-
-        $syncJob->markStep(SyncStep::GitLab);
-        $this->publishProgress(['status' => 'in_progress', 'current_step' => 'gitlab']);
-        $syncService->syncGitLab();
-
-        $syncJob->markStep(SyncStep::Bitrix24);
-        $this->publishProgress(['status' => 'in_progress', 'current_step' => 'bitrix24']);
-        $syncService->syncBitrix24();
-
-        $syncJob->markStep(SyncStep::Matching);
-        $this->publishProgress(['status' => 'in_progress', 'current_step' => 'matching']);
-        $matchAllUnmatched();
-
-        $syncJob->markCompleted();
-        $this->publishProgress(['status' => 'success']);
     }
 
     /**
